+++ tests/Backend/Unit/Console/helper/Greenlight2Room.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Backend\Unit\Console\helper;
 
-class GreenlightRoom
+class Greenlight2Room
 {
     public $id;