+++ tests/Frontend/e2e/AdminSettingsView.cy.js
@@ -55,6 +55,19 @@ describe("Admin settings with edit permission", function () {
           .and("be.disabled");
       });
 
+    cy.get('[data-test="accessibility-statement-url-field"]')
+      .should("be.visible")
+      .and("include.text", "admin.settings.accessibility_statement_url.title")
+      .and(
+        "include.text",
+        "admin.settings.accessibility_statement_url.description",
+      )
+      .within(() => {
+        cy.get("#accessibility-statement-url")
+          .should("have.value", "")
+          .and("be.disabled");
+      });
+
     cy.get('[data-test="pagination-page-size-field"]')
       .should("be.visible")
       .and("include.text", "admin.settings.pagination_page_size.title")