+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Accessibility statement URL setting for the admin UI and footer ([#2873], [#2874])
+- Greenlight v3 import command ([#2664], [#2665])
+
+### Fixed
+
+- Greenlight v2 imported room settings not applied due to disabled expert mode ([#2665])
+
+## [v4.13.0] - 2026-02-23
+
+### Added
+
 - Rate limiting to prevent Room-ID enumeration attacks ([#2518])
 
 ### Changed
@@ -681,6 +692,8 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#2613]: https://github.com/THM-Health/PILOS/pull/2613
 [#2616]: https://github.com/THM-Health/PILOS/pull/2616
 [#2660]: https://github.com/THM-Health/PILOS/issues/2660
+[#2664]: https://github.com/THM-Health/PILOS/issues/2664
+[#2665]: https://github.com/THM-Health/PILOS/pull/2665
 [#2686]: https://github.com/THM-Health/PILOS/pull/2686
 [#2726]: https://github.com/THM-Health/PILOS/pull/2726
 [#2728]: https://github.com/THM-Health/PILOS/pull/2728
@@ -702,7 +715,9 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#2855]: https://github.com/THM-Health/PILOS/pull/2855
 [#2856]: https://github.com/THM-Health/PILOS/issues/2856
 [#2857]: https://github.com/THM-Health/PILOS/pull/2857
-[unreleased]: https://github.com/THM-Health/PILOS/compare/v4.12.0...develop
+[#2873]: https://github.com/THM-Health/PILOS/issues/2873
+[#2874]: https://github.com/THM-Health/PILOS/pull/2874
+[unreleased]: https://github.com/THM-Health/PILOS/compare/v4.13.0...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1
 [v3.0.2]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.2
@@ -725,3 +740,4 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [v4.10.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.10.0
 [v4.11.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.11.0
 [v4.12.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.12.0
+[v4.13.0]: https://github.com/THM-Health/PILOS/releases/tag/v4.13.0